Added user inputs their math question and the result is given

+++ calculator.c
@@ -13,7 +13,7 @@ int multiply(int a, int b) {
     return a * b;
 }
 
-int divide(int a, int b) {
+double divide(int a, int b) {
     if (b == 0) {
         printf("Error: Division by zero\n");
         return 0; // Return 0 or handle error as needed
