+++ calculator.h
@@ -4,6 +4,6 @@
 int add(int a, int b);
 int subtract(int a, int b);
 int multiply(int a, int b);
-int divide(int a, int b);
+double divide(int a, int b);
 
 #endif
